big changes
i realized that my stylesheet it was not good to be responsive, and for that reason i've mantain the comments.
i added new images and the images that not have a path, have the logo image.
the maioraty style that i've change it was for the section called welcomeSection on index.html