work on cart form and buttons related
added eventListener to the buttons on the top of cart page related with the cart length, if the cart is empty, the buttons aren't displayed.
the same happens with inputs on cart form.
The user need  to to choose between delivery or pick up, and if no one are selected or clicked, all the inputs inside form are disabled/readonly